Move merge_ndarray_like_cols to BaseColumnInfo.merge_cols_attributes
