+++ astropy/units/quantity.py
@@ -179,11 +179,10 @@ class QuantityInfo(ParentDtypeInfo):
         col : Quantity (or subclass)
             Empty version of this class consistent with ``cols``
         """
-        from ..table.column import _merge_ndarray_like_cols
 
         # Get merged info attributes like shape, dtype, format, description, etc.
-        attrs = _merge_ndarray_like_cols(cols, metadata_conflicts, name,
-                                         ('meta', 'format', 'description'))
+        attrs = self.merge_cols_attributes(cols, metadata_conflicts, name,
+                                           ('meta', 'format', 'description'))
 
         # Make an empty quantity filled with Nan using the unit of the last one.
         shape = (length,) + attrs.pop('shape')
